test(server): split the multi-project suite before it takes a release
The release of 0.17.0-beta.1 failed on `test timed out after 300000ms`
with `# fail 0`: nothing was red, one file ran out of time and was
cancelled. `multiProjectWorkspaces.test.mjs` had grown to 26 tests, each
starting a real server — 139s here, and past the per-file limit on a
runner already busy from the executables matrix. The two branch CIs before
it had passed on the same file, which is what being at the edge looks
like.

Split by what each half is about. `multiProjectWorkspaces` keeps what one
connection sees: a project's sessions are its own, a client hears about
other projects' activity and not their content, and a file, a session or a
tool call cannot cross between projects. `multiProjectLifecycle` takes the
set of open projects and what the server does to it: opening, closing,
persisting, retiring, pinning, refusing.

What both need — a scriptable RPC server, a second project, the session
seeding that makes one project's history distinguishable — moves to
`multiProjectHarness.mjs`.

Two files get two budgets, and the runner takes them in parallel: 55s and
79s, four to five times under the limit rather than twice. Same 26 tests,
none changed.

server 1,599 passed, nothing skipped, 202s for the whole suite.
